+++ editor/src/com/talosvfx/talos/editor/addons/scene/widgets/directoryview/DirectoryViewWidget.java
@@ -1,37 +1,26 @@
-package com.talosvfx.talos.editor.addons.scene.widgets;
+package com.talosvfx.talos.editor.addons.scene.widgets.directoryview;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Input;
 import com.badlogic.gdx.files.FileHandle;
-import com.badlogic.gdx.graphics.Color;
-import com.badlogic.gdx.graphics.OrthographicCamera;
-import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.g2d.Batch;
-import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.scenes.scene2d.*;
 import com.badlogic.gdx.scenes.scene2d.ui.*;
 import com.badlogic.gdx.scenes.scene2d.utils.*;
 import com.badlogic.gdx.utils.*;
 import com.talosvfx.talos.TalosMain;
 import com.talosvfx.talos.editor.TalosInputProcessor;
-import com.talosvfx.talos.editor.addons.scene.MainRenderer;
 import com.talosvfx.talos.editor.addons.scene.SceneEditorAddon;
-import com.talosvfx.talos.editor.addons.scene.SceneEditorWorkspace;
-import com.talosvfx.talos.editor.addons.scene.assets.AssetRepository;
 import com.talosvfx.talos.editor.addons.scene.assets.GameAsset;
 import com.talosvfx.talos.editor.addons.scene.assets.RawAsset;
 import com.talosvfx.talos.editor.addons.scene.events.PropertyHolderSelected;
-import com.talosvfx.talos.editor.addons.scene.logic.GameObject;
 import com.talosvfx.talos.editor.addons.scene.logic.IPropertyHolder;
 import com.talosvfx.talos.editor.addons.scene.logic.MultiPropertyHolder;
-import com.talosvfx.talos.editor.addons.scene.logic.components.TransformComponent;
 import com.talosvfx.talos.editor.addons.scene.utils.AMetadata;
 import com.talosvfx.talos.editor.addons.scene.utils.importers.AssetImporter;
+import com.talosvfx.talos.editor.addons.scene.widgets.ProjectExplorerWidget;
 import com.talosvfx.talos.editor.notifications.Notifications;
 import com.talosvfx.talos.editor.widgets.ui.ActorCloneable;
-import com.talosvfx.talos.editor.widgets.ui.EditableLabel;
 import com.talosvfx.talos.editor.widgets.ui.FilteredTree;
-import com.talosvfx.talos.editor.widgets.ui.common.ColorLibrary;
 
 import java.io.File;
 import java.io.FileFilter;
@@ -76,11 +65,6 @@ public class DirectoryViewWidget extends Table {
 
             @Override
             public boolean keyDown(InputEvent event, int keycode) {
-                if (selected.size == 1) {
-                    int i = selected.first();
-                    Item item = (Item) items.getChild(i);
-                    item.rename();
-                }
 
                 ProjectExplorerWidget projectExplorer = SceneEditorAddon.get().projectExplorer;
 
@@ -141,9 +125,16 @@ public class DirectoryViewWidget extends Table {
                         selectionStart = -1;
                         reportSelectionChanged();
                     }  else {
-                        Array<FileHandle> current = new Array<>(1);
-                        current.add(fileHandle);
-                        projectExplorer.showContextMenu(current, true);
+                        Gdx.app.postRunnable(new Runnable() {
+                            @Override
+                            public void run () {
+                                Array<FileHandle> current = new Array<>(1);
+                                current.add(fileHandle);
+                                projectExplorer.showContextMenu(current, true);
+                            }
+                        });
+
+                        return true;
                     }
                 } else {
                     if (button == 1) {
@@ -159,8 +150,14 @@ public class DirectoryViewWidget extends Table {
                             selected.add(overItem);
                             reportSelectionChanged();
                         }
-                        Array<FileHandle> selection = convertToFileArray(selected);
-                        projectExplorer.showContextMenu(selection, true);
+                        Gdx.app.postRunnable(new Runnable() {
+                            @Override
+                            public void run () {
+                                Array<FileHandle> selection = convertToFileArray(selected);
+                                projectExplorer.showContextMenu(selection, true);
+                            }
+                        });
+
                     } else if (button == 0) {
                         if (overItem >= 0) {
                             Item item = (Item) items.getChild(overItem);
@@ -218,6 +215,7 @@ public class DirectoryViewWidget extends Table {
         items.pad(20);
         items.wrapSpace(10);
         items.space(10);
+
         ScrollPane scrollPane = new ScrollPane(items);
         scrollPane.setScrollbarsVisible(true);
         Stack stack = new Stack(scrollPane, emptyFolderTable);
@@ -315,6 +313,8 @@ public class DirectoryViewWidget extends Table {
         selected.clear();
         items.clear();
         dragAndDrop.clear();
+
+        //BEGIN DRAG AND DROP
         for (DragAndDrop.Target externalTarget : externalTargets) {
             dragAndDrop.addTarget(externalTarget);
         }
@@ -333,7 +333,7 @@ public class DirectoryViewWidget extends Table {
                 Object object = payload.getObject();
                 if(object instanceof Array) {
 
-                    Array<DirectoryWidgetTrash.ItemView> array = (Array<DirectoryWidgetTrash.ItemView>) object;
+                    Array<Item> array = (Array<Item>) object;
 
                     for (int i = 0; i < selected.size; i++) {
                         int entry = selected.get(i);
@@ -466,6 +466,8 @@ public class DirectoryViewWidget extends Table {
             }
         });
 
+        ////END DRAG AND DROP
+
         FileHandle[] content = directory.list();
         if (content.length == 0) {
             emptyFolderTable.setVisible(true);
@@ -531,6 +533,7 @@ public class DirectoryViewWidget extends Table {
                     if(selected.size == 1) {
                         Item newView = new Item();
                         Actor dragging = ((ActorCloneable) newView).copyActor(item);
+                        dragging.setSize(item.getWidth(), item.getHeight());
                         payload.setDragActor(dragging);
                         if (newView.gameAsset != null) {
                             payload.setObject(newView.gameAsset);
@@ -585,176 +588,6 @@ public class DirectoryViewWidget extends Table {
         }
     }
 
-    /**
-     * Completely ignores the pref size of its children.
-     */
-    private static class ItemGroup extends WidgetGroup {
-        private float prefHeight;
-        private float cellWidth, cellHeight;
-
-        private float space, wrapSpace, padTop, padLeft, padBottom, padRight;
-
-        public ItemGroup () {
-            setTouchable(Touchable.childrenOnly);
-        }
-
-        @Override
-        public void invalidate() {
-            super.invalidate();
-        }
-
-        public void layout () {
-            float padLeft = this.padLeft, padRight = this.padRight, pad = padLeft + padRight, space = this.space, wrapSpace = this.wrapSpace;
-            float width = cellWidth, height = cellHeight;
-            float maxWidth = getWidth() - pad;
-            SnapshotArray<Actor> children = getChildren();
-            int n = children.size;
-            if (n < 1) {
-                return;
-            }
-
-            if (n * width + (n - 1) * space <= maxWidth) { // strategy 1 - align.topLeft and fixed space
-                float rowHeight = height, x = padLeft;
-
-                float startY = getHeight() - padTop - rowHeight;
-                int i = 0, incr = 1;
-                for (; i != n; i += incr) {
-                    Actor child = children.get(i);
-                    Layout layout = null;
-                    if (child instanceof Layout) {
-                        layout = (Layout) child;
-                    }
-                    float y = startY;
-                    y += rowHeight - height;
-                    child.setBounds(x, y, width, height);
-                    x += width + space;
-                    if (layout != null) layout.validate();
-                }
-                prefHeight = padTop + padBottom + cellHeight;
-            } else {
-                float rowY = getHeight() - padTop, groupWidth = getWidth(), xStart = padLeft, x = 0, rowHeight = 0, rowDir = -1;
-
-                int canFit = (int) ((maxWidth - space) / (width + space));
-                if (canFit > 1 && n > 1) { // strategy 2 - align.topCenter and dynamic space between
-                    maxWidth += pad;
-                    space = (maxWidth - canFit * width) / (canFit + 1);
-                    xStart = space;
-
-                    int i = 0, incr = 1;
-                    for (int r = 0; i != n; i += incr) {
-                        Actor child = children.get(i);
-
-                        Layout layout = null;
-                        if (child instanceof Layout) {
-                            layout = (Layout)child;
-                        }
-
-                        if ((i % canFit) == 0 || r == 0) {
-                            x = xStart;
-                            rowHeight = height;
-                            if (r > 0) rowY += wrapSpace * rowDir;
-                            rowY += rowHeight * rowDir;
-                            r += 1;
-                        }
-
-                        float y = rowY;
-                        y += rowHeight - height;
-
-                        child.setBounds(x, y, width, height);
-                        prefHeight = r * height + (r - 1) * wrapSpace + padBottom + padTop;
-
-                        x += width + space;
-
-                        if (layout != null) layout.validate();
-                    }
-                } else { // strategy 3 - one item per row
-                    int i = 0, incr = 1;
-                    for (int r = 0; i != n; i += incr) {
-                        Actor child = children.get(r);
-
-                        Layout layout = null;
-                        if (child instanceof Layout) {
-                            layout = (Layout)child;
-                        }
-
-                        x = xStart;
-                        x += (maxWidth - width) / 2;
-                        rowHeight = height;
-                        if (r > 0) rowY += wrapSpace * rowDir;
-                        rowY += rowHeight * rowDir;
-                        r += 1;
-
-                        float y = rowY;
-                        y += rowHeight - height;
-
-                        child.setBounds(x, y, width, height);
-
-                        if (layout != null) layout.validate();
-                    }
-
-                    prefHeight = padTop + padBottom + n * cellHeight + (n - 1) * wrapSpace;
-                }
-            }
-        }
-
-        /** Sets the padTop, padLeft, padBottom, and padRight to the specified value. */
-        public ItemGroup pad (float pad) {
-            padTop = pad;
-            padLeft = pad;
-            padBottom = pad;
-            padRight = pad;
-            return this;
-        }
-
-        /** Sets the horizontal space between children. */
-        public ItemGroup space (float space) {
-            this.space = space;
-            return this;
-        }
-
-        public void setCellWidth (float width) {
-            this.cellWidth = width;
-        }
-
-        public void setCellHeight (float height) {
-            this.cellHeight = height;
-        }
-
-        public void setCellSize (float size) {
-            setCellSize(size, size);
-        }
-
-        public void setCellSize (float width, float height) {
-            setCellWidth(width);
-            setCellHeight(height);
-        }
-
-        /** Sets the vertical space between rows when wrap is enabled. */
-        public ItemGroup wrapSpace (float wrapSpace) {
-            this.wrapSpace = wrapSpace;
-            return this;
-        }
-
-        protected void drawDebugBounds (ShapeRenderer shapes) {
-            super.drawDebugBounds(shapes);
-            if (!getDebug()) return;
-            shapes.set(ShapeRenderer.ShapeType.Line);
-            if (getStage() != null) shapes.setColor(getStage().getDebugColor());
-            shapes.rect(getX() + padLeft, getY() + padBottom, getOriginX(), getOriginY(), getWidth() - padLeft - padRight,
-                    getHeight() - padBottom - padTop, getScaleX(), getScaleY(), getRotation());
-        }
-
-        @Override
-        public float getMinWidth() {
-            return padLeft + cellWidth + padRight;
-        }
-
-        @Override
-        public float getPrefHeight() {
-            return prefHeight;
-        }
-    }
-
     public FileHandle getCurrentFolder () {
         return fileHandle;
     }
@@ -769,182 +602,4 @@ public class DirectoryViewWidget extends Table {
         dragAndDrop.removeTarget(externalTarget);
     }
 
-    private class Item extends Widget implements ActorCloneable<Item> {
-        private Image icon;
-        private Image brokenStatus;
-        private EditableLabel label;
-
-        private FileHandle fileHandle;
-        private GameAsset<?> gameAsset;
-
-        private GameObject basicGameObject;
-
-        private boolean selected;
-        private boolean mouseover;
-
-        private float padTop = 5, padTopSmall = 2, padTopBig = 7;
-        private float padLeft = 5, padLeftSmall = 2, padLeftBig = 7;
-        private float padBottom = 5, padBottomSmall = 2, padBottomBig = 7;
-        private float padRight = 5, padRightSmall = 2, padRightBig = 7;
-        private float space = 5, spaceSmall = 2, spaceBig = 7;
-
-        public Item () {
-            Skin skin = TalosMain.Instance().getSkin();
-            icon = new Image(null, Scaling.fit, Align.center);
-            brokenStatus = new Image(TalosMain.Instance().getSkin().newDrawable("ic-fileset-file"));
-            brokenStatus.setColor(Color.RED);
-            label = new EditableLabel("text", skin);
-            label.getLabel().setAlignment(Align.center);
-            setTouchable(Touchable.enabled);
-        }
-
-        private void setFile (FileHandle fileHandle) {
-            String name = fileHandle.name();
-            label.setText(name);
-            label.getLabel().setWrap(true);
-            label.getLabel().setEllipsis(true);
-
-            if (fileHandle.isDirectory()) {
-                FileHandle[] content = fileHandle.list();
-                if (content.length == 0) {
-                    icon.setDrawable(TalosMain.Instance().getSkin().getDrawable("ic-folder-big-empty"));
-                } else {
-                    icon.setDrawable(TalosMain.Instance().getSkin().getDrawable("ic-folder-big"));
-                }
-            } else {
-                icon.setDrawable(TalosMain.Instance().getSkin().getDrawable("ic-file-big"));
-                String extension = fileHandle.extension();
-                if (extension.equals("png") || extension.equals("jpg") || extension.equals("jpeg")) {
-                    Texture texture = new Texture(fileHandle);
-                    TextureRegionDrawable drawable = new TextureRegionDrawable(texture);
-                    icon.setDrawable(drawable);
-                    icon.setScaling(Scaling.fit);
-                }
-            }
-
-            this.fileHandle = fileHandle;
-
-            GameAsset<?> assetForPath = AssetRepository.getInstance().getAssetForPath(fileHandle, true);
-            if (assetForPath != null) {
-                gameAsset = assetForPath;
-            }
-
-            if (assetForPath != null) {
-                GameObject parent = new GameObject();
-                parent.addComponent(new TransformComponent());
-                basicGameObject = parent;
-
-                AssetImporter.fromDirectoryView = true; //tom is very naughty dont be like tom
-                boolean success = AssetImporter.createAssetInstance(assetForPath, parent) != null;
-                if (parent.getGameObjects() == null || parent.getGameObjects().size == 0) {
-                    success = false;
-                }
-                AssetImporter.fromDirectoryView = false;
-
-                if (success) {
-                    //Game asset is legit, lets try to make one
-                    GameObject copy = new GameObject();
-                    copy.addComponent(new TransformComponent());
-
-                    AssetImporter.fromDirectoryView = true; //tom is very naughty dont be like tom
-                    AssetImporter.createAssetInstance(assetForPath, copy);
-                    AssetImporter.fromDirectoryView = false;
-
-                    MainRenderer uiSceneRenderer = SceneEditorWorkspace.getInstance().getUISceneRenderer();
-                    Stage stage = SceneEditorWorkspace.getInstance().getStage();
-                    uiSceneRenderer.setCamera((OrthographicCamera)stage.getCamera());
-                    GameObjectActor gameObjectActor = new GameObjectActor(uiSceneRenderer, basicGameObject, copy, true);
-                    gameObjectActor.setFillParent(true);
-                }
-            }
-        }
-
-        @Override
-        public void draw(Batch batch, float parentAlpha) {
-            if (selected) {
-                Drawable bg = ColorLibrary.obtainBackground(TalosMain.Instance().getSkin(), "white", ColorLibrary.BackgroundColor.DARK_GRAY);
-                bg.draw(batch, getX(), getY(), getWidth(), getHeight());
-            } else if (mouseover) {
-                Drawable bg = ColorLibrary.obtainBackground(TalosMain.Instance().getSkin(), "white", ColorLibrary.BackgroundColor.LIGHT_GRAY);
-                bg.draw(batch, getX(), getY(), getWidth(), getHeight());
-            }
-
-            float w = getWidth(), h = getHeight();
-            float padTop, padLeft, padBottom, padRight, space;
-            if (w < 75) {
-                padTop = padTopSmall;
-                padLeft = padLeftSmall;
-                padBottom = padBottomSmall;
-                padRight = padRightSmall;
-                space = spaceSmall;
-            } else if (w >= 75) {
-                padTop = this.padTop;
-                padLeft = this.padLeft;
-                padBottom = this.padBottom;
-                padRight = this.padRight;
-                space = this.space;
-            } else {
-                padTop = padTopBig;
-                padLeft = padLeftBig;
-                padBottom = padBottomBig;
-                padRight = padRightBig;
-                space = spaceBig;
-            }
-
-            w -= padLeft + padRight;
-            h -= padTop + padBottom;
-
-            float lw = w, lh = label.getLabel().getHeight();
-            float lx = getX() + padLeft, ly = getY() + padBottom;
-            label.getLabel().setWidth(lw);
-            label.getTextField().setWidth(lw);
-            label.setPosition(lx, ly);
-            label.draw(batch, parentAlpha);
-
-            float iw = w, ih = h - lh - space - padTop;
-            float ix = getX() + padLeft;
-            float iy = ly + lh + space;
-            icon.setSize(iw, ih);
-            icon.setPosition(ix, iy);
-            icon.draw(batch, parentAlpha);
-
-            if (gameAsset != null && gameAsset.isBroken()) {
-                float bsix = getX() + padLeft, bsiy = iy;
-                brokenStatus.setPosition(bsix, bsiy);
-                brokenStatus.draw(batch, parentAlpha);
-                batch.setColor(Color.WHITE);
-            }
-        }
-
-        @Override
-        public Item copyActor(Item copyFrom) {
-            setFile(copyFrom.fileHandle);
-            return this;
-        }
-
-        public boolean isSelected () {
-            return selected;
-        }
-
-        public void select () {
-            selected = true;
-        }
-
-        public void deselect () {
-            selected = false;
-        }
-
-        public void setMouseover(boolean over) {
-            mouseover = over;
-        }
-
-        public FileHandle getFileHandle() {
-            return fileHandle;
-        }
-
-        public void rename() {
-            label.setStage(getStage());
-            label.setEditMode();
-        }
-    }
 }
